refactor(benchmarks): remove renderer asset facades (#33)
* refactor(benchmarks): remove renderer asset facades

* docs(benchmarks): record canonical asset ownership
