+++ docs/log.md
@@ -2,6 +2,7 @@
 
 ## 2026-08-03
 
+- **Canonical benchmark asset and raster metadata boundaries** — Removed renderer-owned font-loader, preload, baked-artifact, and raster-configuration compatibility facades. Live scenes and finite product/conformance targets now call the discriminated `workloads/font-assets` API directly, while Bitmap atlas, MTSDF extension, and Slug allocation inspection live under `benchmark/low-level/raster`. The complete 316-test gate and production build passed; all 19 isolated conformance/product scenarios remained deterministic, all 42 sequential Presentation cells rendered visible pixels with one renderer, and the retained comparison plus exclusive finite-job recovery probe passed on WebGPU and WebGL2. Live renderer chunks fell again to 9.26/3.38 kB minified/gzip for Bitmap, 7.77/2.85 for MTSDF, and 7.28/2.75 for Slug.
 - **Persistent renderer API cleanup** — Removed the unused standalone Bitmap, MTSDF, and Slug preview constructors and 676 lines of duplicate renderer, RAF, GPU-timer, telemetry, resize, and disposal lifecycle. The remaining contracts are named for persistent scenes, and a boundary regression rejects reintroducing preview entrypoints. The complete 315-test benchmark gate and production build passed; live renderer chunks fell from 16.82/5.19 to 10.37/3.84 kB minified/gzip for Bitmap, 9.02/3.32 to 8.37/3.12 for MTSDF, and 9.26/3.37 to 8.19/3.09 for Slug. All 42 sequential Presentation cells rendered visible pixels with one renderer, and both timed demos completed their authored sequence, returned to Off-axis / 3D, and measured 59.92/60.02 Icon Grid FPS on WebGPU/WebGL2.
 - **Slug conformance scene locality** — Moved the Slug release-role scene definitions from the reusable renderer directory beside the conformance capture that owns and executes them. The capture script and URL-loaded browser probe now resolve the target-owned module directly, and the benchmark boundary regression rejects the obsolete renderer path.
 - **Bitmap conformance target ownership** — Moved the remaining finite Bitmap and source-outline wrappers from the live renderer into `benchmark/targets/conformance/raster/bitmap-capture`, retaining the renderer-neutral finite scene and exact atlas reference under `benchmark/low-level/raster`. Conformance targets, runtime fallback, and the React finite surface now import only the target wrapper; boundary tests reject every old renderer capture path. The complete 314-test benchmark gate and production build passed, and all 19 isolated browser scenarios—including exact Bitmap product and source-outline captures—remained deterministic.
